Issue-184 - Latest CircularEqualizer changes
